Basic file overlaying works, with a few exceptions.
I manually slurped/spit the torquebox config to match the formatting
of immutant's. The overlay doesn't deal well with the webservices
subsystem's endpoint-config elements, because there are two with the
exact same tags and attributes. I think I'm gonna just introduce an
"ignore tag" feature to address that.

I also manually added the enable-welcome-root attribute to the
torquebox file as it seems we're not setting that anymore, and that's
probably a bug? Otherwise, the overlay will result in two vhost
elements with the same name, which probably won't work. Not sure if I
want to get clever about knowing when to replace a node if say, a
subset of its attribute values match.